Split AIS query tests into test_query.py to mirror source layout.
Move query_ais_mxak tests out of test_reader.py so test modules align
with nps_active_space.utils.ais.reader and .query.
